created a new query file for Webinar
