+++ src/database/query/shiksha.ts
@@ -8,7 +8,7 @@ import {
   UpdateCourseRequestPayloadProps,
   UpdateUserChapterInCourseRequestProps,
 } from '@/interfaces';
-import { Course, UserCourse, Webinar } from '@/database';
+import { Course, UserCourse } from '@/database';
 import { modelSelectParams } from '@/constant';
 
 const addACourseToDB = async (
@@ -321,15 +321,6 @@ const getACourseForUserFromDB = async (userId: string, courseId: string) => {
   }
 };
 
-const getAllWebinarsFromDB = async () => {
-  try {
-    const webinars = await Webinar.find();
-    return { data: webinars, error: null };
-  } catch (error) {
-    return { data: null, error };
-  }
-};
-
 export {
   addACourseToDB,
   updateACourseInDB,
@@ -345,5 +336,4 @@ export {
   updateUserCourseChapterInDB,
   getACourseForUserFromDB,
   getAllCourseFromDB,
-  getAllWebinarsFromDB,
 };
