+++ src/database/query/webinar.ts
@@ -1,0 +1,12 @@
+import { Webinar } from '@/database';
+
+const getAllWebinarsFromDB = async () => {
+  try {
+    const webinars = await Webinar.find();
+    return { data: webinars, error: null };
+  } catch (error) {
+    return { data: null, error };
+  }
+};
+
+export { getAllWebinarsFromDB };
